child, pasta: Allow drivers to configure their own interface, let pasta do that
As reported in https://github.com/moby/moby/issues/48257, when Docker
rootless uses pasta through rootlesskit for user-mode connectivity,
IPv6 can't be used for outbound connections because no addresses and
no routes are configured in the container.

The reason is that rootlesskit won't configure IPv6 addresses on the
interface, and at the same time it doesn't ask pasta to do so using
the --config-net option.

Add a ChildDriverInfo() method to childDriver, returning a single
piece of information, that is, whether the driver configures the
network interface by itself, which is true only for pasta. If that's
the case, there's no reason to call activateDev() from setupNet().

Further, in the pasta driver, skip the call to PrepareTap(), because
pasta can take care of that as well.

At the same time, ask pasta to do all that: set up the tap device,
and configure IPv4 and IPv6, using --config-net.

While at it, drop options --no-ra and --no-dhcp, as the container
might want to send router solicitations and DHCP requests even if we
permanently configure IPv4 and IPv6 addresses and routes, and
there's no reason to ignore those requests.

Drop --stderr as well: it doesn't do anything anymore, and it has
been obsoleted in pasta for a while (it will always print to stderr
when starting in foreground anyway).

Link: https://github.com/moby/moby/issues/48257
Signed-off-by: Stefano Brivio <sbrivio@redhat.com>