pasta: Let it run in background, and wait until it forks
pasta(1), like many daemons, will fork to background only once it's
ready to operate (handle traffic). On top of that, it doesn't need an
explicit clean-up phase (unless --no-netns-quit is given), because it
will terminate as soon as the target network namespace goes away.

If we make it run in foreground and proceed as soon as we invoked the
command (in background), we'll claim we're done too early, and tests
such as integration-net.sh will occasionally fail, because pasta is
still initialising sockets and interfaces as UDP packets are sent.

Drop --foreground, and use CombinedOutput() instead of the Start()
method of exec.Command(). Drop the explicit cleanup steps, too.

Signed-off-by: Stefano Brivio <sbrivio@redhat.com>