port/builtin: return proper error for sctp
When trying to forward sctp port we get a useless
"rootlessport spec was not validated" error back even though the spec
was validated before. It is not clear to me if other drivers support
sctp here so I didn't want to change ValidatePortSpec to disallow sctp.

As such simply return a useful error saying the protocol is not
supported.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>